requirements.txt
requests==2.22.0
tqdm==4.32.2
python-telegram-bot==11.1.0
